fix data type bug dan fix debug apk bug

+++ lib/data/api_service.dart
@@ -36,7 +36,7 @@ class ApiService {
     }
   }
 
-  Future<int> getPh() async {
+  Future<double> getPh() async {
     final response = await http.get(Uri.parse("$_baseUrl$_ph"));
     if (response.statusCode == 200) {
       return json.decode(response.body);
